fix two wrong types; update test file; rename identifiers.
